Make Pre-Build step fail-safe. Move BuildInfo into separate file
